Resolved CID 52878 (Argument cannot be negative)

+++ src/world/DBC/DBCStores.h
@@ -944,7 +944,7 @@ class SERVER_DECL DBCStorage
         {
             uint32 i;
             uint32 string_length;
-            long pos;
+            int pos;
 
             FILE* f = fopen(filename, "rb");
             if (f == NULL)
